requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.115.0
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==16.0
reportlab
pypdf
pillow